Refactored program. Added date calculator class and number calculator class Added error checking for operator input
