+++ Calculator2/Calculator2/bin/Debug/net5.0/Calculator2.deps.json
@@ -1,0 +1,23 @@
+{
+  "runtimeTarget": {
+    "name": ".NETCoreApp,Version=v5.0",
+    "signature": ""
+  },
+  "compilationOptions": {},
+  "targets": {
+    ".NETCoreApp,Version=v5.0": {
+      "Calculator2/1.0.0": {
+        "runtime": {
+          "Calculator2.dll": {}
+        }
+      }
+    }
+  },
+  "libraries": {
+    "Calculator2/1.0.0": {
+      "type": "project",
+      "serviceable": false,
+      "sha512": ""
+    }
+  }
+}
