+++ Calculator2/Calculator2/bin/Debug/net5.0/Calculator2.runtimeconfig.dev.json
@@ -1,0 +1,8 @@
+{
+  "runtimeOptions": {
+    "additionalProbingPaths": [
+      "C:\\Users\\Ian.Nkwocha\\.dotnet\\store\\|arch|\\|tfm|",
+      "C:\\Users\\Ian.Nkwocha\\.nuget\\packages"
+    ]
+  }
+}
